Fix module hoisthing problem with buidler
As stated in yarn documentation here:
https://classic.yarnpkg.com/en/docs/workspaces/#toc-limitations-caveats

"The package layout will be different between your workspace and what
your users will get." Unfortunately, 'budler' searches for plugins in
the project root's ./node_modules/ folder, which is not guaranteed to
work with yarn workspaces.

Declaring all buidler dependencies in a root package.json seems to
resolve the issue.